+++ tasks/power/power.py
@@ -127,6 +127,7 @@ class Power:
     @staticmethod
     def wait_fight():
         logger.info(_("进入战斗"))
+        time.sleep(2)
         # for i in range(20):
         #     if auto.find_element("./assets/images/base/not_auto.png", "image", 0.95, crop=(0.0 / 1920, 903.0 / 1080, 144.0 / 1920, 120.0 / 1080)):
         #         logger.info(_("尝试开启自动战斗"))