+++ tasks/activity/__init__.py
@@ -50,6 +50,7 @@ class ActivityManager:
         if not result:
             return []
 
+        logger.debug(result)
         return [box[1][0] for box in result if len(box[1][0]) >= 4]
 
     def _finish(self):