+++ assets/docs/Changelog.md
@@ -2,7 +2,12 @@
 
 ## v1.7.7
 
+### 新功能
+- SMTP 支持发送截图 [#114](https://github.com/moesnow/March7thAssistant/pull/114)
+- 支持 gotify 推送方式 [#112](https://github.com/moesnow/March7thAssistant/pull/112)
 ### 修复
+- 远程桌面多开会错误终止其他用户的游戏进程 [#113](https://github.com/moesnow/March7thAssistant/pull/113) [#35](https://github.com/moesnow/March7thAssistant/issues/35)
+- “异器盈界”文字识别异常
 - 路径中含有英文括号导致解压失败
 - 模拟宇宙完成后无法正常领取奖励